restructure local_instruments in templates to have same structure as customised_instruments and update initialisation_script to import customised and local instruments from the right places
